requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
tensorflow
keras
pyspark
kafka-python
jupyter
